% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMDK.R
\docType{package}
\name{BMDK}
\alias{BMDK}
\title{BMDK: Biomarker Discovery Kit}
\description{
The Biomarker Discovery Kit (BMDK) package provides a framework to aid in bimarker discovery. The GitHub repository also includes a Shiny app for use with the package (see https://github.com/abcsFrederick/BMDK).
}
