% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_bmdk.R
\name{clean_bmdk}
\alias{clean_bmdk}
\title{Clean the BMDK data}
\usage{
clean_bmdk(dat, naThreshold = 0.05, sdMultiplier = 4)
}
\arguments{
\item{dat}{a list containing 3 elements: case, a list of case/control
statuses; feat, a matrix of normalized feature data; maxfeat, a list of max
features from each column in feat}

\item{naThreshold}{Proportion of NAs to allow before dropping a sample or feature due to missingness.}

\item{sdMultiplier}{Number of standard deviations defining the threshold for outlier detection. Any values above sdMultiplier standard deviations away from the mean will be removed as an outlier.}
}
\value{
List containing 3 elements (case, feat, maxfeat) with relevant NA
values and necessary features/samples removed
}
\description{
Identify outliers across samples and across features and set identified
values to NA. Features and samples that contain a threshold ratio of NA
values are removed.
}
