% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers_bmdk.R
\name{detectOutliers_bmdk}
\alias{detectOutliers_bmdk}
\title{Detect outliers in the BMDK data}
\usage{
detectOutliers_bmdk(dat)
}
\arguments{
\item{dat}{a list containing 3 elements: case, a list of case/control statuses;
feat, a matrix of normalized feature data; maxfeat, a list of max features
from each column in feat}
}
\value{
List containing 3 elements (case, feat, maxfeat) with the outliers removed
}
\description{
Using the k Nearest Neighbours Distance method, detect outliers in the data
and remove them. After removing the outliers, renormalize the data.
}
