% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtgini.R
\name{dtgini}
\alias{dtgini}
\title{Runs a Decision Tree Gini Index}
\usage{
dtgini(dat)
}
\arguments{
\item{dat}{a list containing 3 elements: case, a list of case/control statuses;
feat, a matrix of normalized feature data; maxfeat, a list of max features
from each column in feat}
}
\value{
finalresults a list of Gini Indices
}
\description{
This procedure uses each feature to perform a one-node split of all N samples
in the parent node into 2 daughter nodes using N-1 cut points, as in a
decision tree. This procedure's goal is to find the most optimal split to
identify between cases and controls for each feature. To do this, this function
identifies the smallest weighted Gini Index value for each feature.
}
