% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bmdk.R, R/pickFeatures_bmdk.R
\name{filter_bmdk}
\alias{filter_bmdk}
\title{Write documentation here}
\usage{
filter_bmdk(dat)

filter_bmdk(dat)
}
\arguments{
\item{dat}{a list containing 3 elements: case, a list of case/control statuses;
feat, a matrix of normalized feature data; maxfeat, a list of max features
from each column in feat}
}
\value{
NEED TO WRITE

NEED TO WRITE
}
\description{
Write documentation here

Write documentation here
}
