% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knn}
\alias{knn}
\title{Calculate k nearest neighbors}
\usage{
knn(nrow, ncol, f)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{f}{Features matrix}
}
\description{
Calculate k nearest neighbors
}
