% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickFeatures_bmdk.R
\name{pickFeatures_bmdk}
\alias{pickFeatures_bmdk}
\title{Picks the top features to be used in the BMDK classifier}
\usage{
pickFeatures_bmdk(dat)
}
\arguments{
\item{dat}{a list containing 4 elements: case, a list of case/control statuses;
feat, a matrix of normalized feature data; maxfeat, a list of max features
from each column in feat; testresults, a list of statistical test results}
}
\value{
dat a list containing 5 elements: case, a list of case/control statuses;
       feat, a matrix of normalized feature data; maxfeat, a list of max features
       from each column in feat; testresults, a list of statistical test results;
       topfeatures, a list of top features names
}
\description{
Takes the top features from each of the filtering methods and finds the union. Then
runs those features through a Pearson Correlation test and identifies correlated pairs.
Removes the feature with the smaller max value of each pair from the top features list.
}
