% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bmdk.R
\name{read_bmdk}
\alias{read_bmdk}
\title{Read BMDK data}
\usage{
read_bmdk(f)
}
\arguments{
\item{f}{path to input file}
}
\value{
A list with case/control status, a matrix, and max feature value formatted and tidied for BMDK
}
\description{
Read in a test or training dataset with the proper format for BMDK.
}
\examples{
test_data <- read_bmdk(system.file('extdata', 'BMDK_test.txt', package = 'BMDK'))
}
