% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renormalize_bmdk.R
\name{renormalize_bmdk}
\alias{renormalize_bmdk}
\title{Renormalize the BMDK features data}
\usage{
renormalize_bmdk(dat)
}
\arguments{
\item{dat}{a list containing 3 elements: case, a list of case/control statuses;
feat, a matrix of normalized feature data; maxfeat, a list of max features
from each column in feat}
}
\value{
List containing 3 elements (case, feat, maxfeat) renormalized after outliers
        are removed
}
\description{
After outliers are removed, renormalize the features data and store the
new max value of each feature.
}
