% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bmdk.R
\name{filter_bmdk}
\alias{filter_bmdk}
\title{Runs the BMDK features data through a series of filtering methods}
\usage{
filter_bmdk(dat)
}
\arguments{
\item{dat}{a list containing 3 elements: case, a list of case/control statuses;
feat, a matrix of normalized feature data; maxfeat, a list of max features
from each column in feat}
}
\value{
dat a list containing 4 elements: case, a list of case/control statuses;
       feat, a matrix of normalized feature data; maxfeat, a list of max features
       from each column in feat; testresults, a list of statistical test results
}
\description{
Utilizes the Wilcoxon Rank Sum test, the t test, and the Decision Tree Gini
Index to identify the significance of each feature.
}
